+++ examples/persist_session/src/main.rs
@@ -5,13 +5,12 @@ use std::{
 
 use matrix_sdk::{
     config::SyncSettings,
-    matrix_auth::Session,
     room::Room,
     ruma::{
         api::client::filter::FilterDefinition,
         events::room::message::{MessageType, OriginalSyncRoomMessageEvent},
     },
-    Client, Error, LoopCtrl,
+    AuthSession, Client, Error, LoopCtrl,
 };
 use rand::{distributions::Alphanumeric, thread_rng, Rng};
 use serde::{Deserialize, Serialize};
@@ -37,7 +36,7 @@ struct FullSession {
     client_session: ClientSession,
 
     /// The Matrix user session.
-    user_session: Session,
+    user_session: AuthSession,
 
     /// The latest sync token.
     ///
@@ -97,7 +96,7 @@ async fn restore_session(session_file: &Path) -> anyhow::Result<(Client, Option<
         .build()
         .await?;
 
-    println!("Restoring session for {}…", user_session.meta.user_id);
+    println!("Restoring session for {}…", user_session.meta().user_id);
 
     // Restore the Matrix user session.
     client.restore_session(user_session).await?;